refactor(Player): move 로직 Player 로 이전
